sceneManager works
